# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Job Application Tracker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type JobStatus = (typeof JobStatus)[keyof typeof JobStatus];

export const JobStatus = {
  wishlist: "wishlist",
  applied: "applied",
  interviewing: "interviewing",
  offer: "offer",
  rejected: "rejected",
  withdrawn: "withdrawn",
} as const;

export interface Job {
  id: number;
  company: string;
  role: string;
  status: JobStatus;
  location?: string | null;
  url?: string | null;
  salary?: string | null;
  notes?: string | null;
  appliedAt?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateJobRequest {
  company: string;
  role: string;
  status: JobStatus;
  location?: string | null;
  url?: string | null;
  salary?: string | null;
  notes?: string | null;
  appliedAt?: string | null;
}

export interface UpdateJobRequest {
  company?: string;
  role?: string;
  status?: JobStatus;
  location?: string | null;
  url?: string | null;
  salary?: string | null;
  notes?: string | null;
  appliedAt?: string | null;
}

export type JobStatsByStatusItem = {
  status: JobStatus;
  count: number;
};

export interface JobStats {
  total: number;
  byStatus: JobStatsByStatusItem[];
  /** Applications in the last 30 days */
  recentActivity: number;
}

export interface ErrorResponse {
  error: string;
}

export type ListJobsParams = {
  status?: JobStatus;
  search?: string;
};