# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Job Application Tracker API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: jobs
    description: Job application operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List all job applications
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/JobStatus"
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of job applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
    post:
      operationId: createJob
      tags: [jobs]
      summary: Create a new job application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJobRequest"
      responses:
        "201":
          description: Job application created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"

  /jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get a job application by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateJob
      tags: [jobs]
      summary: Update a job application
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateJobRequest"
      responses:
        "200":
          description: Updated job application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteJob
      tags: [jobs]
      summary: Delete a job application
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /jobs/stats:
    get:
      operationId: getJobStats
      tags: [jobs]
      summary: Get dashboard statistics
      responses:
        "200":
          description: Job application statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobStats"

  /jobs/export:
    get:
      operationId: exportJobsCsv
      tags: [jobs]
      summary: Export job applications as CSV
      responses:
        "200":
          description: CSV file
          content:
            text/csv:
              schema:
                type: string

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    JobStatus:
      type: string
      enum:
        - wishlist
        - applied
        - interviewing
        - offer
        - rejected
        - withdrawn

    Job:
      type: object
      properties:
        id:
          type: integer
        company:
          type: string
        role:
          type: string
        status:
          $ref: "#/components/schemas/JobStatus"
        location:
          type: string
          nullable: true
        url:
          type: string
          nullable: true
        salary:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        appliedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - company
        - role
        - status
        - createdAt
        - updatedAt

    CreateJobRequest:
      type: object
      properties:
        company:
          type: string
        role:
          type: string
        status:
          $ref: "#/components/schemas/JobStatus"
        location:
          type: string
          nullable: true
        url:
          type: string
          nullable: true
        salary:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        appliedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - company
        - role
        - status

    UpdateJobRequest:
      type: object
      properties:
        company:
          type: string
        role:
          type: string
        status:
          $ref: "#/components/schemas/JobStatus"
        location:
          type: string
          nullable: true
        url:
          type: string
          nullable: true
        salary:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        appliedAt:
          type: string
          format: date-time
          nullable: true

    JobStats:
      type: object
      properties:
        total:
          type: integer
        byStatus:
          type: array
          items:
            type: object
            properties:
              status:
                $ref: "#/components/schemas/JobStatus"
              count:
                type: integer
            required:
              - status
              - count
        recentActivity:
          type: integer
          description: Applications in the last 30 days
      required:
        - total
        - byStatus
        - recentActivity

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error