# lib/api-zod/src/generated/types/jobStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Job Application Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { JobStatsByStatusItem } from "./jobStatsByStatusItem";

export interface JobStats {
  total: number;
  byStatus: JobStatsByStatusItem[];
  /** Applications in the last 30 days */
  recentActivity: number;
}