# lib/api-zod/src/generated/types/jobStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Job Application Tracker API
 * OpenAPI spec version: 0.1.0
 */

export type JobStatus = (typeof JobStatus)[keyof typeof JobStatus];

export const JobStatus = {
  wishlist: "wishlist",
  applied: "applied",
  interviewing: "interviewing",
  offer: "offer",
  rejected: "rejected",
  withdrawn: "withdrawn",
} as const;