# lib/api-zod/src/generated/types/updateJobRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Job Application Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { JobStatus } from "./jobStatus";

export interface UpdateJobRequest {
  company?: string;
  role?: string;
  status?: JobStatus;
  location?: string | null;
  url?: string | null;
  salary?: string | null;
  notes?: string | null;
  appliedAt?: Date | null;
}